Refactor: #106 Group auth and user utils
