+++ src/Services/auth.js
@@ -1,0 +1,5 @@
+import { LOCAL_STORAGE } from 'utils/constants';
+
+export const logout = () => {
+  LOCAL_STORAGE.remove('token');
+};
